+++ src/components/FileTree/TreeViewNode.svelte
@@ -42,7 +42,6 @@
   export let icon = undefined;
 
   import { afterUpdate, getContext } from "svelte";
-  import { addFileTab } from "../Content/Editor/scripts/Tab";
   import RenameModel from "../Modal/RenameModel.svelte";
   import LeafNodeMenu from "./LeafNodeMenu.svelte";
